+++ java-backend/src/main/java/org/kframework/backend/java/symbolic/BuiltinFunction.java
@@ -14,13 +14,15 @@ import org.kframework.utils.errorsystem.KExceptionManager;
 import org.kframework.utils.inject.Builtins;
 
 import java.lang.invoke.MethodHandle;
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.MethodType;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.google.common.collect.ImmutableSet;
 import com.google.inject.Inject;
+import com.google.inject.Provider;
 
 /**
  * Utility class that handles the builtin (hooked) operations and their Java
@@ -31,23 +33,6 @@ import com.google.inject.Inject;
 public class BuiltinFunction {
 
     /**
-     * Set of hook module names excluded from evaluation during compilation, when each rule's
-     * right-hand side and condition are partially evaluated. Certain functions, like functions
-     * performing I/O operations or meta operations should only be evaluated at runtime.
-     *
-     * @deprecated Use dummy hook implementations that throw {@link ImpureFunctionException} instead.
-     * @see org.kframework.backend.java.symbolic.KILtoBackendJavaKILTransformer#evaluateDefinition(org.kframework.backend.java.kil.Definition)
-     * @see org.kframework.backend.java.symbolic.KILtoBackendJavaKILTransformer#evaluateRule(org.kframework.backend.java.kil.Rule, org.kframework.backend.java.kil.Definition)
-     */
-    @Deprecated
-    private final ImmutableSet<String> hookMetaModules = ImmutableSet.of(
-            "#META-K",
-            "MetaK",
-            "Visitor",
-            "#FRESH",
-            "Substitution");
-
-    /**
      * Map of {@link KLabelConstant} representation of builtin (hooked) operations to
      * {@link Method} representation of Java implementation of said operations.
      */
@@ -55,12 +40,27 @@ public class BuiltinFunction {
 
 
     /**
+     * Constructs a builtin function table mapping KLabels to the methods that implement their builtin
+     * functions.
      *
-     * @param definition
-     * @param injector
+     * The "impure" attribute on productions is used to exclude functions from evaluation during compilation,
+     * when each rule's right-hand side and condition are partially evaluated. Certain functions, like functions
+     * performing I/O operations or meta operations should only be evaluated at runtime.
+     *
+     * @see org.kframework.backend.java.symbolic.KILtoBackendJavaKILTransformer#evaluateDefinition(org.kframework.backend.java.kil.Definition)
+     * @see org.kframework.backend.java.symbolic.KILtoBackendJavaKILTransformer#evaluateRule(org.kframework.backend.java.kil.Rule, org.kframework.backend.java.kil.Definition)
      */
     @Inject
-    public BuiltinFunction(Context context, @Builtins Map<String, MethodHandle> hookProvider, KExceptionManager kem, Tool tool) {
+    public BuiltinFunction(Context context, @Builtins Map<String, Provider<MethodHandle>> hookProvider, KExceptionManager kem, Tool tool) {
+        MethodHandles.Lookup lookup = MethodHandles.lookup();
+        MethodType hookType = MethodType.methodType(Term.class, Object[].class);
+        MethodHandle throwImpureExceptionHandle;
+        try {
+            throwImpureExceptionHandle = lookup.findStatic(BuiltinFunction.class,
+                    "throwImpureException", hookType);
+        } catch (NoSuchMethodException | IllegalAccessException e) {
+            throw KExceptionManager.internalError("Failed to load partial evaluation hook implementation", e);
+        }
         for (String label : context.klabels.keySet()) {
             for (Production production : context.productionsOf(label)) {
                 if (production.getKLabel().equals(label) // make sure the label is a Klabel
@@ -71,7 +71,8 @@ public class BuiltinFunction {
                      * exclude hook from evaluation during compilation if the hook is dynamic
                      * in nature (is related to I/O or to meta properties).
                      * */
-                    if (tool == Tool.KOMPILE && hookMetaModules.contains(hookPrefix)) {
+                    if (tool == Tool.KOMPILE && production.containsAttribute("impure")) {
+                        table.put(KLabelConstant.of(label, context), throwImpureExceptionHandle);
                         continue;
                     }
 
@@ -83,10 +84,14 @@ public class BuiltinFunction {
                         continue;
                     }
 
-                    table.put(KLabelConstant.of(label, context), hookProvider.get(hookAttribute));
+                    table.put(KLabelConstant.of(label, context), hookProvider.get(hookAttribute).get());
                 }
             }
         }
+    }
+
+    private static Term throwImpureException(Object... args) {
+        throw new ImpureFunctionException();
     }
 
     /**
