+++ k-distribution/include/builtins/fresh.k
@@ -1,6 +1,6 @@
 // Copyright (c) 2014 K Team. All Rights Reserved.
 
 module FRESH-HOOKS
-  syntax KItem ::= fresh(String)   [function, hook(#FRESH:fresh)]
+  syntax KItem ::= fresh(String)   [function, hook(#FRESH:fresh), impure]
 endmodule
