+++ k-distribution/include/modules/k-functional-visitor.k
@@ -19,7 +19,7 @@ module K-FUNCTIONAL-VISITOR
    * if IfLabel(K',,IFList) holds then replace K' with VisitLabel(K',,VisitList)
    * otherwise recurse on the children of K'.
    */
-  syntax KItem ::= "#visit" "(" K "," KLabel "," KList "," KLabel "," KList ")"   [function, hook(Visitor:visit)]
+  syntax KItem ::= "#visit" "(" K "," KLabel "," KList "," KLabel "," KList ")"   [function, hook(Visitor:visit), impure]
 
   /*
    * Particular forms of #visit (default argument KList is .KList, default
