+++ k-distribution/samples/java_rewrite_engine/tutorial/1_k/5_types/meta-k.k
@@ -1,6 +1,6 @@
 // Copyright (c) 2013-2014 K Team. All Rights Reserved.
 module META-K
-  syntax Set ::= #variables(KItem)        [function, hook(#META-K:#variables)]
-  syntax KItem ::= #rename(KItem, Set)    [function, hook(#META-K:#rename)]
+  syntax Set ::= #variables(KItem)        [function, hook(#META-K:#variables), impure]
+  syntax KItem ::= #rename(KItem, Set)    [function, hook(#META-K:#rename), impure]
 endmodule
