+++ k-distribution/samples/java_rewrite_engine/tutorial/2_languages/4_logik/logik-desired2.k
@@ -36,7 +36,7 @@ module LOGIK-DESIRED2
        <clause> forall _ . Ls -> L </clause>
     [transition]
 
-  syntax KItem ::= variables(K) [function, hook(MetaK:variables)]
-  syntax KItem ::= variablesMap(K) [function, hook(MetaK:variablesMap)]
-  syntax KItem ::= renameVariables(K) [function, hook(MetaK:renameVariables)]
+  syntax KItem ::= variables(K) [function, hook(MetaK:variables), impure]
+  syntax KItem ::= variablesMap(K) [function, hook(MetaK:variablesMap), impure]
+  syntax KItem ::= renameVariables(K) [function, hook(MetaK:renameVariables), impure]
 endmodule
