+++ k-distribution/tutorial/1_k/5_types/lesson_9.5/meta-k.k
@@ -1,5 +1,5 @@
 // Copyright (c) 2013-2014 K Team. All Rights Reserved.
 module META-K
-  syntax Set ::= #variables(K)    [function, hook(#META-K:#variables)]
-  syntax K ::= #rename(K, Set)    [function, hook(#META-K:#rename)]
+  syntax Set ::= #variables(K)    [function, hook(#META-K:#variables), impure]
+  syntax K ::= #rename(K, Set)    [function, hook(#META-K:#rename), impure]
 endmodule
